fix: move abstraction to reset enabled/handler
